review correction: mock Visit and renaming
